Fixed loading bug when coming from character selector
DnDBeyond updated to change the character selector so that hitting "view" now loads the character sheet without actually changing pages, just paths, which meant the script wouldn't load. It worked fine otherwise when linking directly to a character, or refreshing after selection. This fixes it so the base script will also load on the character selector, and then track the page change once a character is selected and still function properly.